+++ src/components/comment/CommentInputModal.module.scss
@@ -7,7 +7,6 @@
   display: flex;
   align-items: flex-end;
   justify-content: center;
-  padding: 0 1.25rem;
 
   .dialog {
     position: relative;
